refactor: add parser generated bindings facade
