+++ src/parser/src/generated-bindings.js
@@ -1,0 +1,2 @@
+export { default as GameMakerLanguageParserListenerBase } from "../generated/GameMakerLanguageParserListener.js";
+export { default as GameMakerLanguageParserVisitorBase } from "../generated/GameMakerLanguageParserVisitor.js";
